requirements.txt
Package               Version
--------------------- ---------
autocorrect           2.5.0
autopep8              1.5.7
click                 8.0.1
imutils               0.5.4
joblib                1.0.1
nltk                  3.6.3
numpy                 1.21.1
opencv-contrib-python 4.5.3.56
opencv-python         3.4.8.29
Pillow                8.3.2
pip                   21.2.4
PyAudio               0.2.11
pycodestyle           2.7.0
PyQt5                 5.15.4
PyQt5-Qt5             5.15.2
PyQt5-sip             12.9.0
pyspellchecker        0.6.2
regex                 2021.8.28
setuptools            57.1.0
textblob              0.15.3
toml                  0.10.2
tqdm                  4.62.3
wheel                 0.36.2
wordsegment           1.3.1
